requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.4; python_version >= '3.6'
django-crispy-forms==1.12.0
django==3.2
pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1; python_version >= '3.5'
typing-extensions==3.10.0.0; python_version < '3.8'
